third_party/WebKit/LayoutTests/editing/inserting/insert_h6_to_h6.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => assert_selection(
    '<div contenteditable><h6>foo|</h6></div>',
    'insertHTML <h6>bar</h6>',
    '<div contenteditable><h6>foobar|</h6></div>'),
    'insert H6 to H6');
</script>
